Updated for new refactored codebase
